# ServerSideNode.js/dist/api.yml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /user/favoriteGames:
    get:
      tags:
      - user
      summary: ' Returns the favorite games of the user'
      description: Find  favorite games according to the user
      operationId: getFavoriteGames
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/futureGame'
        "204":
          description: There is no favorite games
  /user/max3FavoriteGames:
    get:
      tags:
      - user
      summary: ' Returns max 3 favorite games of the user'
      description: Find  favorite games according to the user and return max 3 of them
      operationId: getMax3FavoriteGames
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/futureGame'
        "204":
          description: There is no favorite games
  /user/favoritePlayers:
    get:
      tags:
      - user
      summary: ' Returns the favorite players of the user'
      description: Find  favorite players according to the user
      operationId: getFavoritePlayers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/playerPreview'
        "404":
          description: there is no favorite games
  /user/favoriteTeams:
    get:
      tags:
      - user
      summary: ' Returns the favorite teams of the user'
      description: Find  favorite teams according to the user
      operationId: getFavoriteTeams
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/team'
        "404":
          description: there is no favorite teams
...
  /user/addFavoritePlayer:
    post:
      tags:
      - user
      summary: mark a player as a favorite
      description: add a new favorite player to the user's favorites
      operationId: addFavoritePlayer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
      responses:
        "200":
          description: success
        "409":
          description: this player already tags as favorite
  /user/deleteFavoritePlayer:
    post:
      tags:
      - user
      summary: delete a player from  the favorites
      description: delete favorite player from   user's favorites
      operationId: deleteFavoritePlayer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
      responses:
        "200":
          description: success
        "409":
          description: this player isnot exist in  favorites
  /user/addFavoriteTeam:
    post:
      tags:
      - user
      summary: mark a team as a favorite
      description: add a new favorite team to the user's favorites
      operationId: addFavoriteTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
      responses:
        "200":
          description: success
        "409":
          description: this team already tags as favorite
  /user/addFavoriteGame:
    post:
      tags:
      - user
      summary: mark a game as a favorite
      description: add a new favorite game to the user's favorites
      operationId: addFavoriteGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_2'
      responses:
        "200":
          description: success
        "409":
          description: this game already tags as favorite
  /user/deleteFavoriteGame:
    post:
      tags:
      - user
      summary: delete a game as a favorite
      description: remobe game from  user's favorites
      operationId: deleteFavoriteGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_2'
      responses:
        "200":
          description: success
        "409":
          description: the games isnt exist in favorites
...
components:
  schemas:
...
    playerPreview:
      type: object
      properties:
        fullName:
          type: string
          example: Cristiano Ronaldo
        groupName:
          type: string
          example: Juventus
        picture:
          type: string
          format: url
          example: https://www.google.com/search?q=ronaldo&tbm=isch&ved=2ahUKEwiZ553y8s3wAhVSgc4BHYgYA5gQ2-cCegQIABAA&oq=ronaldo&gs_lcp=CgNpbWcQAzIFCAAQsQMyAggAMgIIADIICAAQsQMQgwEyAggAMgIIADICCAAyAggAMgIIADICCAA6BAgAEAM6BAgAEEM6BwgAELEDEENQzKMMWMe-DGCWwQxoAHAAeACAAaMCiAHUCpIBBTAuNi4xmAEAoAEBqgELZ3dzLXdpei1pbWfAAQE&sclient=img&ei=GeegYNn2CNKCur4PiLGMwAk&bih=553&biw=1280&rlz=1C1CHZL_enIL833IL833&safe=active#imgrc=B1-_nkTDWaykrM
        position:
          type: number
          example: 7
...
    futureGame:
      type: object
      properties:
        date:
          type: string
          format: date
        time:
          type: string
          format: time
          example: 18:30
        home_team:
          type: string
          example: Real Madrid
        away_team:
          type: string
          example: Barcelona
        stadium:
          type: string
          example: Camp NOU
...
    team:
      type: object
      properties:
        name:
          type: string
          example: Real Madrid
        logo:
          type: string
          format: url
          example: https://www.google.com/search?q=real+madrid&tbm=isch&ved=2ahUKEwi-44DZ9M3wAhUM0oUKHVwgB1oQ2-cCegQIABAA&oq=real+madrid&gs_lcp=CgNpbWcQAzIHCAAQsQMQQzIECAAQQzICCAAyAggAMgIIADICCAAyAggAMgIIADICCAAyAggAOgUIABCxAzoECAAQA1C5oDdYw7k3YPq6N2gAcAB4AIABzAGIAY4PkgEGMC4xMC4xmAEAoAEBqgELZ3dzLXdpei1pbWfAAQE&sclient=img&ei=_eigYP7wBoyklwTcwJzQBQ&bih=553&biw=1280&rlz=1C1CHZL_enIL833IL833&safe=active#imgrc=1y-wf1Vu9pHFrM
...
    body:
      properties:
        playerId:
          type: integer
    body_1:
      properties:
        teamId:
          type: integer
    body_2:
      properties:
        gameId:
          type: number
...